Moved specs and implementation of record destruction
When a node is being removed due to its securable being removed, sometimes the
securable instance can be returned as a parent of the immediate children
records while re-parenting, even if it is about to be removed later in the
same request.  This securable instance has no node anymore in the
after_destroy callback.  Then, the logic of re-parentig was moved to the
securable model, and the implementation now seems to be more robust, since in
the after_destroy callback we're sure that both the node and the securable
instance were removed, and the children securables will not return the
securable as a parent.