Record destruction
Addressed issue when removing a record which destroys its dependants.  Also
added spec about what happens with child records that are not re-parented, that
is, child records with one parent that are not destroyed or are not dependants
from the parent record destroyed (thus their `#parents` method returns an
empty array).