Security manager during action execution
All the request/response cycle can access the SecurityManager instance, even
if rendering from inside a `rescue_from`.  The manager is set when the request
processing starts, and is removed right before it finishes.  This makes
exception handling happen inside the `run_with_security_manager` loop, so the
security manager is available even when rescuing exceptions.