Made Parenter's logic more general
By using the recently created 'Inheritance.recognizes?' method, it was
possible to generalize the logic on Parenter (more specifically, for the
ancestors-related methods) and make it more bendable to changes on the
notion of a "Inheritance-Aware record". For instance, it was possible to
remove the need for explicit exceptions for the GlobalRecord, using the
fact that the Inheritance module doesn't recognize it.