AccessControl.Node now checks for responsivity to .ac_node
This has changed to allow for an analogue implementation of .Principal method,
which in turn will check for responsivity to .ac_principal.

The old behaviour was checking for inclusion of Securable, which has no
analogue in Principal's miracle world.  The closes thing there is
ActiveRecordSubject, which is really analogue to ActiveRecordSecurable.
Changing .Node to check for ActiveRecordSecurable seemed fair at first, but
wasn't good enough as it would cause us to include that module in GlobalNode,
to make it work through .Node.  GlobalNode isn't a ActiveRecord model, so
there's no point in doing that inclusion.  The simplest solution is simply
check for .ac_node.