Improved (further) AssignmentCombination algorithm and code
Due to an Array-based lookup, the previous algorithm was taking a
O(n^2) time to perform the initial instantiations of the assignments.
Now, due to a mix of Sets and Hashes, plus memoization where convenient,
the algorithm is two orders of magnitude faster for a number of items in
the order of 50.000.
As a bonus, some code cleaning was made.