Query restriction will now be done with a JOIN instead of a WHERE condition
We could use a subquery, just making the small modifications to use the
conditions as such, but MySQL doesn't help...  That's why we plan to use a
join, which will be done over a view (probably one view per securable type),
which in turn is a select in a real table which doesn't exist yet (grouping
the ids to avoid repetition) and this table will be somehow kept in sync with
the rest of the system.

Simple, isn't it?