Reverting back to using conditions
We are starting to use subqueries whenever they are considered the best
solutions.  If a database doesn't support subqueries, or its support is
deficient (as MySQL now in major Linux distros), that database is not
supported by this gem anymore.

Simply put: this gem now requires good subquery support from the database.

We studied many solutions for the problem of querying records without exposing
those that the current principal cannot see.  Clearly, the best solution is
using a subquery in a "cache" table (which is yet to be created).  Subqueries
simplify the overall query, have good performance when its support is decent
and are easier to maintain.  Other solutions are overly complex, impose
maintenance problems, and/or don't have good performance/scalability.

As an alternative to MySQL, there is MariaDB, which is a drop-in replacement
for it and has good support for subqueries starting from version 5.3.x.