Removed the 'stub soup' from ActiveRecordSecurable specs
By making the module configurable, it was possible to remove many of the
"global" stubs, that were many times unrelated to the behaviour under
test. By toggling only the relevant functionality, it is possible to
test the module without the need to add "global" stubs to maintain the
previous tests.

Also, the configuration might be useful to users of the library that,
for some reason, don't want to activate all the functionality.