Use datasets in place of static lists as much as possible
Also, rely on the Sequel adapter being capable of providing the first id
inserted in bulk inserts and on the ids being consecutive.

This is generally the case in MySQL with InnoDB tables, when the option
innodb_autoinc_lock_mode is left at its default value (1 --
"consecutive") or changed to 0 ("traditional").

The other possible value for the option is 2 (interleaved), which makes
the ids possibly interleave under concurrency, and render our
assumptions false.

Thus, AccessControl now requires a database configured to always
generate ids in consecutive order for a given bulk insert and a Sequel
adapter for the database capable to retrieve the id of the first row
inserted.