Fix role destruction
Depending on the order role assignments are destroyed, a stale assignment may
have its #destroy method issued causing a database error in Sequel.  This
happens because real assignments automatically destroy their child (effective)
assignments.

Because this is a bit hard to test (due to order dependency, it would
require some sort of dependency injection in the API to inject assignments in
the order it could provoke the error), tests for this fix were omitted.