Use prepended module for method protection
The following use case will not be protected anymore:

  class Model < Framework::Model
    include AccessControl::Securable
    protect :method, with: 'some permission'
    def method
      'do something'
    end
  end

  class SubModel < Model
    def method
      'do something without calling super'
    end
  end

If permission protection is needed in SubModel#method, then it should call
`super'.  Otherwise it'll bypass permission checks.  This was previously
supported, but the code required to use `alias_method' twice and had to be
injected on initialization (or allocation) of the class.  Not very clean...

Then, I choose to sacrifice this use case (requiring this to modestly call
`super' at any sensible code path) in order to make things neater, and
possibly increase performance since no singleton_class must be accessed only
to provide method protection anymore.