Reverting association unrestriction
Now, no association is unrestricted by default, as previously (in v0.19.x).
But now `belongs_to' associations will behave like `has_one' with respect to
throwing Unauthorized: they don't.

`has_one' associations work as `has_many', which has already been filtered by
usual listing restrictions, except that they return only the first record
found.  In other words, when all records are filtered out, it'll be equivalent
to `[].first', which is always `nil'.

When only some, but not all, of the records are filtered out, we had, and
still have, a problem: depending on the permissions of the current user, some
records may sometimes be filtered out, sometimes not, changing the `.first'
one in the collection (and ordering issues may apply).  This behavior from
v0.19.x is back, and it'll be up to the application to ensure its `has_one's
are all well behaved, that is, either return *one* single record (maybe by
ensuring uniqueness on the key in the reflected table through the use of a
database index?) or *no* record at all, thus making the `.first' record either
the only available one, or `nil' through means of unauthorization or
unavailability.